move all of cutils.c into cutils.h because all the definitions are static
