+++ libregexp.c
@@ -28,7 +28,7 @@
 #include <string.h>
 #include <assert.h>
 
-#include "cutils.c"
+#include "cutils.h"
 #include "libregexp.h"
 
 #if defined(__sun)
