+++ quickjs.c
@@ -42,7 +42,7 @@
 #include <time.h>
 #include <math.h>
 
-#include "cutils.c"
+#include "cutils.h"
 #include "list.h"
 #include "quickjs.h"
 #include "libregexp.h"
