+++ unicode_gen.c
@@ -31,7 +31,7 @@
 #include <ctype.h>
 #include <time.h>
 
-#include "cutils.c"
+#include "cutils.h"
 
 /* define it to be able to test unicode.c */
 //#define USE_TEST
